requirements.txt
Flask
pyyaml
psycopg2-binary